support-portal/backend/requirements.txt
chalice==1.30.0
boto3==1.35.36
python-dotenv==1.0.1
psycopg2-binary==2.9.9

